% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\docType{data}
\name{bcr_pattern}
\alias{bcr_pattern}
\title{Regular expression for BCR genes}
\format{
Character string regex.
}
\usage{
bcr_pattern
}
\description{
Regular expression for BCR genes
}
\keyword{datasets}
