% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biowrenchR-package.R
\docType{package}
\name{biowrenchR-package}
\alias{biowrenchR}
\alias{biowrenchR-package}
\title{biowrenchR: Data Structure Pipelines for Genomic Analysis}
\description{
Helpers for genomic file formatting, human-mouse gene mapping, gene-set
filtering, isoform summarization, annotation joins, and plotting utilities.
}
\section{Source layout}{

\describe{
\item{\code{R/formats.R}}{List/data-frame helpers; GMT and GCT I/O}
\item{\code{R/biomart.R}}{Load human/mouse/joint BioMart dictionaries}
\item{\code{R/convert.R}}{Ortholog conversion for vectors, data frames, expression}
\item{\code{R/genesets.R}}{Gene-set helpers and family regex patterns}
\item{\code{R/annotate.R}}{Annotation joins}
\item{\code{R/summarize.R}}{Isoform-to-gene expression aggregation}
\item{\code{R/plot-palette.R}}{Color palettes and \code{get_palette()}}
\item{\code{R/plot-export.R}}{PDF plot export helpers}
\item{\code{R/plot-strips.R}}{Facet strip recoloring}
\item{\code{R/internal.R}}{Cache, extdata readers, validators}
}
}

\section{Format I/O}{

\code{read_gmt()}, \code{write_gmt()}, \code{read_gct()}, \code{write_gct()},
\code{list_to_df()}, \code{df_to_list()}
}

\section{Conversion}{

\code{get_conversion_dict()}, \code{convert_genes()}, \code{convert_df()}, \code{convert_exprs()}
}

\section{BioMart loaders}{

\code{load_biomart()}, \code{load_biomart_human()}, \code{load_biomart_mouse()}
}

\section{Gene-set helpers}{

\code{xy_genes()}, \code{mt_genes()}, \code{cell_cycle_genes()}, \code{genes_by_pattern()},
and pattern constants (\code{bcr_pattern}, \code{tcr_pattern}, ...)
}

\section{Annotation and summarization}{

\code{annotate_genes()}, \code{summarize_genes()}
}

\section{Plotting helpers}{

\code{get_palette()}, \code{palette_list()}, \code{display_palettes()},
\code{save_plot()}, \code{save_jupyter_plot()}, \code{add_strip_pal()}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/}
  \item Report bugs at \url{https://github.com/}
}

}
\author{
\strong{Maintainer}: Matthew Hung \email{matthew.hung@crick.ac.uk}

}
\keyword{internal}
