% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-palette.R
\name{display_palettes}
\alias{display_palettes}
\title{Display built-in palettes}
\usage{
display_palettes(n = NULL)
}
\arguments{
\item{n}{Optional max colors to show per palette. Default shows all.}
}
\value{
A ggplot2 object.
}
\description{
Display built-in palettes
}
\examples{
\dontrun{
display_palettes()
display_palettes(n = 5)
}

}
