% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\docType{data}
\name{hb_pattern}
\alias{hb_pattern}
\title{Regular expression for hemoglobin genes}
\format{
Character string regex.
}
\usage{
hb_pattern
}
\description{
Regular expression for hemoglobin genes
}
\keyword{datasets}
