% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{list_to_df}
\alias{list_to_df}
\title{Convert a list of vectors to a data frame}
\usage{
list_to_df(x)
}
\arguments{
\item{x}{A list of vectors (typically character).}
}
\value{
A data frame with one column per list element.
}
\description{
Pads shorter vectors with empty strings so all columns share the same length.
}
\examples{
my_list <- list(a = c("A", "B", "C"), b = c("D", "E"))
list_to_df(my_list)

}
