% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\docType{data}
\name{mhc_pattern}
\alias{mhc_pattern}
\title{Regular expression for MHC genes}
\format{
Character string regex.
}
\usage{
mhc_pattern
}
\description{
Regular expression for MHC genes
}
\keyword{datasets}
