% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{mt_genes}
\alias{mt_genes}
\title{Mitochondrial genes}
\usage{
mt_genes(org, ...)
}
\arguments{
\item{org}{Organism to query (\code{"human"} or \code{"mouse"}).}

\item{...}{Additional arguments passed to \code{\link[=load_biomart_human]{load_biomart_human()}} /
\code{\link[=load_biomart_mouse]{load_biomart_mouse()}}.}
}
\value{
Character vector of mitochondrial gene symbols.
}
\description{
Mitochondrial genes
}
\examples{
\dontrun{
mt_genes("human")
}

}
