% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\docType{data}
\name{mt_pattern}
\alias{mt_pattern}
\title{Regular expression for mitochondrially encoded gene symbols (\verb{MT-} / \verb{mt-})}
\format{
Character string regex.
}
\usage{
mt_pattern
}
\description{
Regular expression for mitochondrially encoded gene symbols (\verb{MT-} / \verb{mt-})
}
\keyword{datasets}
