% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\docType{data}
\name{rb_pattern}
\alias{rb_pattern}
\title{Regular expression for ribosomal protein genes}
\format{
Character string regex.
}
\usage{
rb_pattern
}
\description{
Regular expression for ribosomal protein genes
}
\keyword{datasets}
