% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\docType{data}
\name{tcr_pattern}
\alias{tcr_pattern}
\title{Regular expression for TCR genes}
\format{
Character string regex.
}
\usage{
tcr_pattern
}
\description{
Regular expression for TCR genes
}
\keyword{datasets}
