% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{xy_genes}
\alias{xy_genes}
\title{Genes on X and Y chromosomes}
\usage{
xy_genes(org, ...)
}
\arguments{
\item{org}{Organism to query (\code{"human"} or \code{"mouse"}).}

\item{...}{Additional arguments passed to \code{\link[=load_biomart_human]{load_biomart_human()}} /
\code{\link[=load_biomart_mouse]{load_biomart_mouse()}}.}
}
\value{
Character vector of gene symbols on chromosomes X and Y.
}
\description{
Genes on X and Y chromosomes
}
\examples{
\dontrun{
xy_genes("human")
}

}
